requirements.txt
pyaml
jinja2
pytest